Separate infrastructure entities into dedicated files
